[misc] Fix `FixedBitSet` output `Array[Byte]` size
- Fix the `FixedBitSet` output `Array[Byte]` size to be 8-byte aligned so that
it is compatible with Arrow validity buffers

- Move `aveobench` benchmarks into `com.nec.cyclone.benchmarks` package
