+++ aveobench/src/main/scala/com/nec/cyclone/benchmarks/CompressedVeTransferBenchmarks.scala
@@ -1,6 +1,7 @@
-package com.nec.cyclone
+package com.nec.cyclone.benchmarks
 
-import com.nec.cyclone.colvector._
+import com.nec.cyclone.colvector.BytePointerColVectorOps._
+import com.nec.cyclone.colvector.CompressedBytePointerColVector
 import com.nec.colvector._
 import com.nec.colvector.SeqOptTConversions._
 import com.nec.ve.VeProcess.OriginalCallingContext.Automatic._
@@ -12,7 +13,7 @@ import org.bytedeco.veoffload.{veo_proc_handle, veo_thr_ctxt}
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.infra.Blackhole
 
-object CompressedBatchVeTransferBenchmarks {
+object CompressedVeTransferBenchmarks {
   implicit val source = VeColVectorSource(getClass.getName)
 
   @State(Scope.Benchmark)
@@ -31,10 +32,10 @@ object CompressedBatchVeTransferBenchmarks {
     var handle: veo_proc_handle = _
     var process: VeProcess.WrappingVeo = _
     var tcontext: veo_thr_ctxt = _
-    var input: CompressedBytePointerColBatch = _
+    var inputs: Seq[CompressedBytePointerColVector] = _
 
     // Temporary data
-    val data = MBuf.empty[CompressedVeColBatch]
+    val data = MBuf.empty[VeColVector]
 
     @Setup(Level.Trial)
     def setup0: Unit = {
@@ -42,8 +43,8 @@ object CompressedBatchVeTransferBenchmarks {
       tcontext = veo.veo_context_open(handle)
       process = VeProcess.WrappingVeo(handle, tcontext, source, VeProcessMetrics.noOp)
 
-      // Initialize CompressedBytePointerColBatch
-      input = BytePointerColBatch(InputSamples.bpcv(typ, ncolumns, size)).compressed
+      // Initialize CompressedBytePointerColVector
+      inputs = InputSamples.bpcv(typ, ncolumns, size).map(_.compressed)
     }
 
     @TearDown(Level.Invocation)
@@ -60,18 +61,35 @@ object CompressedBatchVeTransferBenchmarks {
   }
 }
 
-class CompressedBatchVeTransferBenchmarks {
+class MultipleCompressedVeTransferBenchmarks {
   @Benchmark
   @Fork(value = 1)
-  @Warmup(iterations = 5)
-  @Measurement(iterations = 10)
+  @Warmup(iterations = 2)
+  @Measurement(iterations = 5)
   @BenchmarkMode(Array(Mode.AverageTime))
-  def benchmark1(fixture: CompressedBatchVeTransferBenchmarks.Fixture): Int = {
+  def benchmark1(fixture: CompressedVeTransferBenchmarks.Fixture): Int = {
     implicit val process = fixture.process
     implicit val source = process.source
     implicit val metrics = process.veProcessMetrics
 
-    fixture.data += fixture.input.toCompressedVeColBatch
-    fixture.input.columns.size
+    fixture.data ++= fixture.inputs.map(_.toVeColVector)
+    fixture.inputs.size
+  }
+
+  @Benchmark
+  @Fork(value = 1)
+  @Warmup(iterations = 2)
+  @Measurement(iterations = 5)
+  @BenchmarkMode(Array(Mode.AverageTime))
+  def benchmark2(fixture: CompressedVeTransferBenchmarks.Fixture): Int = {
+    implicit val process = fixture.process
+    implicit val source = process.source
+    implicit val metrics = process.veProcessMetrics
+
+    // Allocate
+    val resultFs = fixture.inputs.map(_.asyncToVeColVector)
+    // Transfer asynchronously and wait
+    fixture.data ++= resultFs.map(_.apply()).map(_.get())
+    fixture.inputs.size
   }
 }
