+++ src/test/scala/com/nec/colvector/SparkSqlColumnVectorConversionsUnitSpec.scala
@@ -32,7 +32,7 @@ final class SparkSqlColumnVectorConversionsUnitSpec extends AnyWordSpec {
 
     // Check validity buffer
     val validityBuffer = colvec.buffers(1)
-    validityBuffer.capacity() should be ((input.size / 8.0).ceil.toLong)
+    validityBuffer.capacity() should be ((input.size / 64.0).ceil.toLong * 8)
     val bitset = FixedBitSet.from(validityBuffer)
     0.until(input.size).foreach(i => bitset.get(i) should be (input(i).nonEmpty))
